Only print mmap address at boot.
Turns out GRUB can only guarentee that after handing control to the kernel, we will be in 32bit protected mode and that some of the registers will have certain values.
Nothing can be guarenteed about any other data returned by GRUB or the bootloader. Not to worry we won't need it until later.